requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alpaca-py==0.42.0
alpaca-trade-api==3.2.0
altair==5.5.0
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
deprecation==2.1.0
fonttools==4.59.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
hmmlearn==0.3.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keyboard==0.13.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
msgpack==1.0.3
multidict==6.6.3
narwhals==2.0.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
patsy==1.0.1
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sseclient-py==1.8.0
statsmodels==0.14.5
streamlit==1.47.1
ta==0.11.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
watchdog==6.0.0
websocket-client==1.8.0
websockets==10.4
yarl==1.20.1
